0.4.0 release✨
