+++ .gitignore
@@ -1,0 +1,3 @@
+.DS_Store
+__pycache__
+*.db
