+++ vesta/__version__.py
@@ -1,4 +1,4 @@
 __title__ = 'vesta'
 __description__ = 'simple gpu monitoring script'
 __url__ = 'https://github.com/a-maumau/vesta'
-__version__ = '0.3.0'
+__version__ = '0.4.0'
