+++ vesta/path_util.py
