+++ vesta/static/css/main.css
@@ -76,6 +76,13 @@ body {
 }
 /**********************************************/
 
+/* top ****************************************/
+.navbar-brand #vesta_ver{
+    padding-left: 15px;
+}
+
+/**********************************************/
+
 /* collapsible content setting ****************/
 .collapsible {
     color: #444;
